Level 5 - Naive Rag Project/requirements.txt
streamlit
langchain_groq 
langchain_community 
pypdf 
sentence_transformers
python-dotenv
chromadb
langchain-chroma
# transformers
langchain-text-splitters
rank_bm25
langchain-huggingface
# langchain-pinecone